fucntion moved from pipelinebuilder  to buildpipelineconfig
